+++ juniper_codegen/src/impl_scalar.rs
@@ -6,7 +6,9 @@ use quote::quote;
 
 #[derive(Debug)]
 struct ScalarCodegenInput {
-    ident: Option<syn::Ident>,
+    impl_for_type: Option<syn::PathSegment>,
+    custom_data_type: Option<syn::PathSegment>,
+    custom_data_type_is_struct: bool,
     resolve_body: Option<syn::Block>,
     from_input_value_arg: Option<syn::Ident>,
     from_input_value_body: Option<syn::Block>,
@@ -16,9 +18,85 @@ struct ScalarCodegenInput {
     from_str_result: Option<syn::Type>,
 }
 
+fn get_first_method_arg(
+    inputs: syn::punctuated::Punctuated<syn::FnArg, syn::Token![,]>,
+) -> Option<syn::Ident> {
+    if let Some(fn_arg) = inputs.first() {
+        match fn_arg {
+            syn::FnArg::Typed(pat_type) => match &*pat_type.pat {
+                syn::Pat::Ident(pat_ident) => return Some(pat_ident.ident.clone()),
+                _ => (),
+            },
+            _ => (),
+        }
+    }
+
+    None
+}
+
+fn get_method_return_type(output: syn::ReturnType) -> Option<syn::Type> {
+    match output {
+        syn::ReturnType::Type(_, return_type) => Some(*return_type),
+        _ => None,
+    }
+}
+
+// Find the enum type by inspecting the type parameter on the return value
+fn get_enum_type(return_type: &Option<syn::Type>) -> Option<syn::PathSegment> {
+    if let Some(return_type) = return_type {
+        match return_type {
+            syn::Type::Path(type_path) => {
+                let path_segment = type_path
+                    .path
+                    .segments
+                    .iter()
+                    .find(|ps| match ps.arguments {
+                        syn::PathArguments::AngleBracketed(_) => true,
+                        _ => false,
+                    });
+
+                if let Some(path_segment) = path_segment {
+                    match &path_segment.arguments {
+                        syn::PathArguments::AngleBracketed(generic_args) => {
+                            let generic_type_arg =
+                                generic_args.args.iter().find(|generic_type_arg| {
+                                    match generic_type_arg {
+                                        syn::GenericArgument::Type(_) => true,
+                                        _ => false,
+                                    }
+                                });
+
+                            if let Some(generic_type_arg) = generic_type_arg {
+                                match generic_type_arg {
+                                    syn::GenericArgument::Type(the_type) => match the_type {
+                                        syn::Type::Path(type_path) => {
+                                            if let Some(path_segment) =
+                                                type_path.path.segments.first()
+                                            {
+                                                return Some(path_segment.clone());
+                                            }
+                                        }
+                                        _ => (),
+                                    },
+                                    _ => (),
+                                }
+                            }
+                        }
+                        _ => (),
+                    }
+                }
+            }
+            _ => (),
+        }
+    }
+
+    None
+}
+
 impl syn::parse::Parse for ScalarCodegenInput {
     fn parse(input: syn::parse::ParseStream) -> syn::parse::Result<Self> {
-        let mut ident: Option<syn::Ident> = None;
+        let mut impl_for_type: Option<syn::PathSegment> = None;
+        let mut enum_data_type: Option<syn::PathSegment> = None;
         let mut resolve_body: Option<syn::Block> = None;
         let mut from_input_value_arg: Option<syn::Ident> = None;
         let mut from_input_value_body: Option<syn::Block> = None;
@@ -28,14 +106,17 @@ impl syn::parse::Parse for ScalarCodegenInput {
         let mut from_str_result: Option<syn::Type> = None;
 
         let parse_custom_scalar_value_impl: syn::ItemImpl = input.parse()?;
+        // To implement a custom scalar for a struct, it's required to
+        // specify a generic type and a type bound
+        let custom_data_type_is_struct: bool =
+            !parse_custom_scalar_value_impl.generics.params.is_empty();
 
         match *parse_custom_scalar_value_impl.self_ty {
-            syn::Type::Path(type_path) => match type_path.path.segments.first() {
-                Some(path_segment) => {
-                    ident = Some(path_segment.ident.clone());
+            syn::Type::Path(type_path) => {
+                if let Some(path_segment) = type_path.path.segments.first() {
+                    impl_for_type = Some(path_segment.clone());
                 }
-                _ => (),
-            },
+            }
             _ => (),
         }
 
@@ -46,47 +127,16 @@ impl syn::parse::Parse for ScalarCodegenInput {
                         resolve_body = Some(method.block);
                     }
                     "from_input_value" => {
-                        match method.sig.inputs.first() {
-                            Some(fn_arg) => match fn_arg {
-                                syn::FnArg::Typed(pat_type) => match &*pat_type.pat {
-                                    syn::Pat::Ident(pat_ident) => {
-                                        from_input_value_arg = Some(pat_ident.ident.clone())
-                                    }
-                                    _ => (),
-                                },
-                                _ => (),
-                            },
-                            _ => (),
-                        }
-
-                        match method.sig.output {
-                            syn::ReturnType::Type(_, return_type) => {
-                                from_input_value_result = Some(*return_type);
-                            }
-                            _ => (),
-                        }
-
+                        from_input_value_arg = get_first_method_arg(method.sig.inputs);
+                        from_input_value_result = get_method_return_type(method.sig.output);
                         from_input_value_body = Some(method.block);
                     }
                     "from_str" => {
-                        match method.sig.inputs.first() {
-                            Some(fn_arg) => match fn_arg {
-                                syn::FnArg::Typed(pat_type) => match &*pat_type.pat {
-                                    syn::Pat::Ident(pat_ident) => {
-                                        from_str_arg = Some(pat_ident.ident.clone())
-                                    }
-                                    _ => (),
-                                },
-                                _ => (),
-                            },
-                            _ => (),
-                        }
+                        from_str_arg = get_first_method_arg(method.sig.inputs);
+                        from_str_result = get_method_return_type(method.sig.output);
 
-                        match method.sig.output {
-                            syn::ReturnType::Type(_, return_type) => {
-                                from_str_result = Some(*return_type);
-                            }
-                            _ => (),
+                        if !custom_data_type_is_struct {
+                            enum_data_type = get_enum_type(&from_str_result);
                         }
 
                         from_str_body = Some(method.block);
@@ -97,8 +147,16 @@ impl syn::parse::Parse for ScalarCodegenInput {
             };
         }
 
+        let custom_data_type = if custom_data_type_is_struct {
+            impl_for_type.clone()
+        } else {
+            enum_data_type
+        };
+
         Ok(ScalarCodegenInput {
-            ident,
+            impl_for_type,
+            custom_data_type,
+            custom_data_type_is_struct,
             resolve_body,
             from_input_value_arg,
             from_input_value_body,
@@ -120,44 +178,81 @@ pub fn build_scalar(attributes: TokenStream, body: TokenStream, is_internal: boo
     };
 
     let input = syn::parse_macro_input!(body as ScalarCodegenInput);
-    let ident = input.ident.unwrap();
-    let resolve_body = input.resolve_body.unwrap();
-    let from_input_value_arg = input.from_input_value_arg.unwrap();
-    let from_input_value_body = input.from_input_value_body.unwrap();
-    let from_input_value_result = input.from_input_value_result.unwrap();
-    let from_str_arg = input.from_str_arg.unwrap();
-    let from_str_body = input.from_str_body.unwrap();
-    let from_str_result = input.from_str_result.unwrap();
 
-    // TODO: Code below copied from derive_scalar_value.rs#impl_scalar_struct. REFACTOR!
+    let impl_for_type = input
+        .impl_for_type
+        .expect("Unable to find target for implementation target for `GraphQLScalar`");
+    let custom_data_type = input
+        .custom_data_type
+        .expect("Unable to find custom scalar data type");
+    let resolve_body = input
+        .resolve_body
+        .expect("Unable to find body of `resolve` method");
+    let from_input_value_arg = input
+        .from_input_value_arg
+        .expect("Unable to find argument for `from_input_value` method");
+    let from_input_value_body = input
+        .from_input_value_body
+        .expect("Unable to find body of `from_input_value` method");
+    let from_input_value_result = input
+        .from_input_value_result
+        .expect("Unable to find return type of `from_input_value` method");
+    let from_str_arg = input
+        .from_str_arg
+        .expect("Unable to find argument for `from_str` method");
+    let from_str_body = input
+        .from_str_body
+        .expect("Unable to find body of `from_str` method");
+    let from_str_result = input
+        .from_str_result
+        .expect("Unable to find return type of `from_str` method");
 
-    let name = attrs.name.unwrap_or_else(|| ident.to_string());
-
-    let crate_name = if is_internal {
-        quote!(crate)
-    } else {
-        quote!(juniper)
+    let name = attrs
+        .name
+        .unwrap_or_else(|| impl_for_type.ident.to_string());
+    let crate_name = match is_internal {
+        true => quote!(crate),
+        _ => quote!(juniper),
     };
-
     let description = match attrs.description {
-        Some(val) => quote!( .description( #val ) ),
+        Some(val) => quote!(.description(#val)),
         None => quote!(),
+    };
+    let async_generic_type = match input.custom_data_type_is_struct {
+        true => quote!(__S),
+        _ => quote!(#custom_data_type),
+    };
+    let async_generic_type_decl = match input.custom_data_type_is_struct {
+        true => quote!(<#async_generic_type>),
+        _ => quote!(),
+    };
+    let generic_type = match input.custom_data_type_is_struct {
+        true => quote!(S),
+        _ => quote!(#custom_data_type),
+    };
+    let generic_type_decl = match input.custom_data_type_is_struct {
+        true => quote!(<#generic_type>),
+        _ => quote!(),
+    };
+    let generic_type_bound = match input.custom_data_type_is_struct {
+        true => quote!(where #generic_type: #crate_name::ScalarValue,),
+        _ => quote!(),
     };
 
     let _async = quote!(
-        impl<__S> ::#crate_name::GraphQLTypeAsync<__S> for #ident
+        impl#async_generic_type_decl #crate_name::GraphQLTypeAsync<#async_generic_type> for #impl_for_type
         where
-            __S: #crate_name::ScalarValue + Send + Sync,
-            Self: #crate_name::GraphQLType<__S> + Send + Sync,
+            #async_generic_type: #crate_name::ScalarValue + Send + Sync,
+            Self: #crate_name::GraphQLType<#async_generic_type> + Send + Sync,
             Self::Context: Send + Sync,
             Self::TypeInfo: Send + Sync,
         {
             fn resolve_async<'a>(
                 &'a self,
                 info: &'a Self::TypeInfo,
-                selection_set: Option<&'a [#crate_name::Selection<__S>]>,
-                executor: &'a #crate_name::Executor<Self::Context, __S>,
-            ) -> #crate_name::BoxFuture<'a, #crate_name::ExecutionResult<__S>> {
+                selection_set: Option<&'a [#crate_name::Selection<#async_generic_type>]>,
+                executor: &'a #crate_name::Executor<Self::Context, #async_generic_type>,
+            ) -> #crate_name::BoxFuture<'a, #crate_name::ExecutionResult<#async_generic_type>> {
                 use #crate_name::GraphQLType;
                 use futures::future;
                 let v = self.resolve(info, selection_set, executor);
@@ -169,9 +264,8 @@ pub fn build_scalar(attributes: TokenStream, body: TokenStream, is_internal: boo
     quote!(
         #_async
 
-        impl<S> #crate_name::GraphQLType<S> for #ident
-        where
-            S: #crate_name::ScalarValue,
+        impl#generic_type_decl #crate_name::GraphQLType<#generic_type> for #impl_for_type
+        #generic_type_bound
         {
             type Context = ();
             type TypeInfo = ();
@@ -182,10 +276,10 @@ pub fn build_scalar(attributes: TokenStream, body: TokenStream, is_internal: boo
 
             fn meta<'r>(
                 info: &Self::TypeInfo,
-                registry: &mut #crate_name::Registry<'r, S>,
-            ) -> #crate_name::meta::MetaType<'r, S>
+                registry: &mut #crate_name::Registry<'r, #generic_type>,
+            ) -> #crate_name::meta::MetaType<'r, #generic_type>
             where
-                S: 'r,
+                #generic_type: 'r,
             {
                 registry.build_scalar_type::<Self>(info)
                     #description
@@ -195,39 +289,36 @@ pub fn build_scalar(attributes: TokenStream, body: TokenStream, is_internal: boo
             fn resolve(
                 &self,
                 info: &(),
-                selection: Option<&[#crate_name::Selection<S>]>,
-                executor: &#crate_name::Executor<Self::Context, S>,
-            ) -> #crate_name::ExecutionResult<S> {
-                #crate_name::GraphQLType::resolve(&self.0, info, selection, executor)
+                selection: Option<&[#crate_name::Selection<#generic_type>]>,
+                executor: &#crate_name::Executor<Self::Context, #generic_type>,
+            ) -> #crate_name::ExecutionResult<#generic_type> {
+                Ok(#resolve_body)
             }
         }
 
-        impl<S> #crate_name::ToInputValue<S> for #ident
-        where
-            S: #crate_name::ScalarValue,
+        impl#generic_type_decl #crate_name::ToInputValue<#generic_type> for #impl_for_type
+        #generic_type_bound
         {
-            fn to_input_value(&self) -> #crate_name::InputValue<S> {
+            fn to_input_value(&self) -> #crate_name::InputValue<#generic_type> {
                 let v = #resolve_body;
                 #crate_name::ToInputValue::to_input_value(&v)
             }
         }
 
-        impl<S> #crate_name::FromInputValue<S> for #ident
-        where
-            S: #crate_name::ScalarValue,
+        impl#generic_type_decl #crate_name::FromInputValue<#generic_type> for #impl_for_type
+        #generic_type_bound
         {
-            fn from_input_value(#from_input_value_arg: &#crate_name::InputValue<S>) -> #from_input_value_result {
+            fn from_input_value(#from_input_value_arg: &#crate_name::InputValue<#generic_type>) -> #from_input_value_result {
                 #from_input_value_body
             }
         }
 
-        impl<S> #crate_name::ParseScalarValue<S> for #ident
-        where
-            S: #crate_name::ScalarValue,
-        {
-            fn from_str<'a>(
-                #from_str_arg: #crate_name::parser::ScalarToken<'a>,
-            ) -> #from_str_result {
+        impl#generic_type_decl #crate_name::ParseScalarValue<#generic_type> for #impl_for_type
+        #generic_type_bound
+            {
+                fn from_str<'a>(
+                    #from_str_arg: #crate_name::parser::ScalarToken<'a>,
+                ) -> #from_str_result {
                 #from_str_body
             }
         }