+++ juniper/CHANGELOG.md
@@ -24,6 +24,8 @@ See [#569](https://github.com/graphql-rust/juniper/pull/569).
 
 - remove old `graphql_object!` macro, rename `object` proc macro to `graphql_object`
 
+- remove old `graphql_scalar!` macro, rename `scalar` proc macro to `graphql_scalar`
+
 - Remove deprecated `ScalarValue` custom derive (renamed to GraphQLScalarValue)
 
 - `graphql_union!` macro removed, replaced by `#[graphql_union]` proc macro